requirements.txt
Django==1.10.3
MySQL-python==1.2.5
Pillow==2.8.1
Scrapy==0.24.6
Twisted==15.2.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.5.1
billiard==3.3.0.20
celery==3.1.18
cffi==1.7.0
characteristic==14.3.0
contextlib2==0.5.4
coverage==4.2
cryptography==1.4
cssselect==0.9.1
django-bootstrap-pagination==1.6.2
django-bootstrap3==8.0.0
django-braces==1.9.0
-e git://github.com/alons45/django-pagination.git@f32d675bcc4c42408095c7fec36e84bc77c451a0#egg=django_pagination-master
django-pagseguro2==1.1.1
django-waffle==0.11.1
enum34==1.0.4
google-api-python-client==1.4.0
gunicorn==19.3.0
httplib2==0.9.1
httpretty==0.8.6
idna==2.0
ipaddress==1.0.7
kombu==3.0.35
lxml==3.4.4
mechanize==0.2.5
mixpanel==4.3.1
oauth2client==1.4.11
pep8==1.7.0
pyOpenSSL==0.15.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.13
pyflakes==1.3.0
python-dateutil==2.5.3
python-memcached==1.54
pytz==2015.4
queuelib==1.2.2
raven==5.27.0
redis==2.10.3
requests==2.7.0
rsa==3.1.4
selenium==2.45.0
service-identity==14.0.0
simplejson==3.7.2
six==1.9.0
sorl-thumbnail==12.4a1
sqlparse==0.1.16
suds==0.4
uritemplate==0.6
w3lib==1.11.0
wsgiref==0.1.2
xmltodict==0.10.2
zope.interface==4.1.2
